django-react-projects/Crud/backend/pyproject.toml
[tool.poetry]
name = "django-react-project-crud"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-ckeditor = "^6.7.1"
django-cors-headers = "^4.4.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
